Maintainability: created a BoundingBox superclass for YOLOWriter and VOCWriter
