+++ libs/boundingBox.py
@@ -1,0 +1,10 @@
+
+class BoundingBox(object):
+    def __init__(self):
+        self.boxlist = []
+
+    def addBndBox(self, xmin, ymin, xmax, ymax, name, difficult):
+        bndbox = {'xmin': xmin, 'ymin': ymin, 'xmax': xmax, 'ymax': ymax}
+        bndbox['name'] = name
+        bndbox['difficult'] = difficult
+        self.boxlist.append(bndbox)
