+++ libs/hashableQListWidgetItem.py
@@ -5,6 +5,7 @@ from PyQt5.QtGui import *
 from PyQt5.QtCore import *
 from PyQt5.QtWidgets import *
 
+
 class HashableQListWidgetItem(QListWidgetItem):
 
     def __init__(self, *args):
